+++ auditlog_tests/manage.py
@@ -3,7 +3,7 @@ import os
 import sys
 
 if __name__ == "__main__":
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "auditlog_tests.test_settings")
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "test_settings")
 
     from django.core.management import execute_from_command_line
