+++ auditlog_tests/test_app/fixtures/m2m_test_fixture.json
@@ -1,6 +1,6 @@
 [
   {
-    "model": "auditlog_tests.manyrelatedmodel",
+    "model": "test_app.manyrelatedmodel",
     "pk": 1,
     "fields": {
       "recursive": [1],
@@ -8,7 +8,7 @@
     }
   },
   {
-    "model": "auditlog_tests.manyrelatedothermodel",
+    "model": "test_app.manyrelatedothermodel",
     "pk": 1,
     "fields": {}
   }
