+++ auditlog_tests/test_app/urls.py
@@ -1,13 +1,13 @@
 from django.contrib import admin
 from django.urls import path
 
-from auditlog_tests.views import SimpleModelDetailview
+from .views import SimpleModelDetailView
 
 urlpatterns = [
     path("admin/", admin.site.urls),
     path(
         "simplemodel/<int:pk>/",
-        SimpleModelDetailview.as_view(),
+        SimpleModelDetailView.as_view(),
         name="simplemodel-detail",
     ),
 ]
