+++ tox.ini
@@ -10,8 +10,9 @@ envlist =
 [testenv]
 setenv =
   COVERAGE_FILE={toxworkdir}/.coverage.{envname}
+changedir = auditlog_tests
 commands =
-    coverage run --source auditlog runtests.py
+    coverage run --source auditlog ./manage.py test
     coverage xml
 deps =
     django42: Django>=4.2,<4.3
